3º Curso (2021-2022)/1º Cuatrimestre/PSI/P3/requirements.txt
coverage==5.5
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.6
django-allauth==0.45.0
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.1
Faker==8.11.0
flake8==3.9.2
image==1.5.33
Pillow==8.3.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.1.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
static3==0.7.0
text-unidecode==1.3
urllib3==1.26.6
zipp==3.5.0
dj-database-url==0.5.0
gunicorn==20.0.4
whitenoise==5.2.0